requirements.txt
mlserver==1.3.1
Pillow==9.5.0
pydantic==1.10.7
Requests==2.30.0
uvicorn==0.22.0
